GUI Lessons: grid()
Added notes on the grid() method as a better option than the pack()
method from tkinter. The new script, named gui2.py, is the recreation of
the sample code found in Murach's book. A modification was also added to
the mushroom_profit.py script that changes the profit entry widget to
read-only.